Autodetect image type when adding a disk via a custom path
When adding a disk via a custom path the code used to default to a raw
image type although a user could add qcow2 images. Instead of adding a
image type select element, auto detect the image type with qemu-img and
limit the available executing/memory usage to circumvent overusage by
potential mailicious images. If an image has a backing file a warning is
shown as we cannot currently import such images. This requires
specifying the backing files image type to be set by the user which will
be done in a future PR.

Fixes #639